requirements.txt
click==7.1.2
Flask==1.1.2
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.22
numpy==1.19.2
pandas==1.1.3
protobuf==3.13.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
WTForms==2.3.3
